+++ src/nodeData.test.ts
@@ -1,4 +1,9 @@
-import { nodeChangersByTeam, UserStatsAccumulator } from "./nodeData";
+import {
+  NO_TEAM_SYMBOL,
+  nodeChangersByTeam,
+  topTeamsPartitioned,
+  UserStatsAccumulator,
+} from "./nodeData";
 import { FileNode, GitData, LocData, NodeLayout } from "./polyglot_data.types";
 import { UserAliases, UserTeams } from "./state";
 
@@ -257,5 +262,126 @@ describe("aggregating node info by team", () => {
       ["teamB", { commits: 1, lines: 2, days: new Set([1]), files: 1 }],
     ]);
     expect(nodeChangers!).toEqual(expected);
+  });
+});
+
+function testTeamStat(
+  teamname: string,
+  commits: number
+): [name: string, stats: UserStatsAccumulator] {
+  return [teamname, { commits, lines: 1, days: new Set(), files: 1 }];
+}
+
+describe("finding top teams as partitions", () => {
+  test("returns single team if only one team returned", () => {
+    const stats: Map<string, UserStatsAccumulator> = new Map([
+      testTeamStat("foo", 1),
+    ]);
+    const partitioned: string[] | undefined = topTeamsPartitioned(
+      stats,
+      "commits",
+      3,
+      true
+    );
+    expect(partitioned!).toEqual(["foo", "foo", "foo"]);
+  });
+  test("returns empty if no teams have data", () => {
+    // empty map
+    expect(topTeamsPartitioned(new Map(), "commits", 3, true)).toBeUndefined();
+    // no stats in map
+    const stats: Map<string, UserStatsAccumulator> = new Map([
+      testTeamStat("foo", 0),
+      testTeamStat("bar", 0),
+    ]);
+    const partitioned: string[] | undefined = topTeamsPartitioned(
+      stats,
+      "commits",
+      3,
+      true
+    );
+    expect(partitioned).toBeUndefined();
+  });
+  test("returns three teams in alphabetical order if stats split evenly", () => {
+    const stats: Map<string, UserStatsAccumulator> = new Map([
+      testTeamStat("foo", 1),
+      testTeamStat("baz", 1),
+      testTeamStat("bar", 1),
+    ]);
+    const partitioned: string[] | undefined = topTeamsPartitioned(
+      stats,
+      "commits",
+      3,
+      true
+    );
+    expect(partitioned!).toEqual(["bar", "baz", "foo"]);
+  });
+  test("returns team with 67% of total twice", () => {
+    const stats: Map<string, UserStatsAccumulator> = new Map([
+      testTeamStat("foo", 67),
+      testTeamStat("bar", 32),
+      testTeamStat("baz", 1),
+    ]);
+    const partitioned: string[] | undefined = topTeamsPartitioned(
+      stats,
+      "commits",
+      3,
+      true
+    );
+    expect(partitioned!).toEqual(["bar", "foo", "foo"]);
+  });
+  test("won't include teams with less than 1/6 of total", () => {
+    const stats: Map<string, UserStatsAccumulator> = new Map([
+      testTeamStat("foo", 9),
+      testTeamStat("baz", 1),
+      testTeamStat("bat", 1),
+      testTeamStat("bag", 1),
+    ]);
+    const partitioned: string[] | undefined = topTeamsPartitioned(
+      stats,
+      "commits",
+      3,
+      true
+    );
+    expect(partitioned!).toEqual(["foo", "foo"]);
+  });
+  test("will include teams with quota of 1/6 of total", () => {
+    const stats: Map<string, UserStatsAccumulator> = new Map([
+      testTeamStat("foo", 10),
+      testTeamStat("baz", 1),
+      testTeamStat("bat", 1),
+    ]);
+    const partitioned: string[] | undefined = topTeamsPartitioned(
+      stats,
+      "commits",
+      3,
+      true
+    );
+    expect(partitioned!).toEqual(["foo", "foo", "foo"]);
+  });
+  test("won't include NO_TEAM team if not wanted, though NO_TEAM stats used in totalling", () => {
+    const stats: Map<string, UserStatsAccumulator> = new Map([
+      testTeamStat("foo", 9),
+      testTeamStat(NO_TEAM_SYMBOL, 3),
+    ]);
+    const partitioned: string[] | undefined = topTeamsPartitioned(
+      stats,
+      "commits",
+      3,
+      false
+    );
+    expect(partitioned!).toEqual(["foo", "foo"]);
+  });
+  test("will include NO_TEAM team requested", () => {
+    const stats: Map<string, UserStatsAccumulator> = new Map([
+      testTeamStat("foo", 9),
+      testTeamStat(NO_TEAM_SYMBOL, 3),
+    ]);
+    const partitioned: string[] | undefined = topTeamsPartitioned(
+      stats,
+      "commits",
+      3,
+      true
+    );
+    expect(partitioned!).toEqual([NO_TEAM_SYMBOL, "foo", "foo"]);
   });
 });